put factory functions in their own file (for replaceability)
