+++ BUILD
@@ -1583,6 +1583,7 @@ grpc_cc_library(
         "src/core/lib/debug/stats_data.cc",
         "src/core/lib/event_engine/endpoint_config.cc",
         "src/core/lib/event_engine/event_engine.cc",
+        "src/core/lib/event_engine/event_engine_factory.cc",
         "src/core/lib/event_engine/sockaddr.cc",
         "src/core/lib/http/format_request.cc",
         "src/core/lib/http/httpcli.cc",
@@ -1744,7 +1745,7 @@ grpc_cc_library(
         "src/core/lib/debug/stats.h",
         "src/core/lib/debug/stats_data.h",
         "src/core/lib/event_engine/endpoint_config_internal.h",
-        "src/core/lib/event_engine/event_engine_internal.h",
+        "src/core/lib/event_engine/event_engine_factory.h",
         "src/core/lib/event_engine/sockaddr.h",
         "src/core/lib/http/format_request.h",
         "src/core/lib/http/httpcli.h",
