+++ gRPC-Core.podspec
@@ -870,7 +870,7 @@ Pod::Spec.new do |s|
                       'src/core/lib/event_engine/endpoint_config.cc',
                       'src/core/lib/event_engine/endpoint_config_internal.h',
                       'src/core/lib/event_engine/event_engine.cc',
-                      'src/core/lib/event_engine/event_engine_internal.h',
+                      'src/core/lib/event_engine/event_engine_factory.h',
                       'src/core/lib/event_engine/sockaddr.cc',
                       'src/core/lib/event_engine/sockaddr.h',
                       'src/core/lib/gpr/alloc.cc',
@@ -1790,7 +1790,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/debug/stats_data.h',
                               'src/core/lib/debug/trace.h',
                               'src/core/lib/event_engine/endpoint_config_internal.h',
-                              'src/core/lib/event_engine/event_engine_internal.h',
+                              'src/core/lib/event_engine/event_engine_factory.h',
                               'src/core/lib/event_engine/sockaddr.h',
                               'src/core/lib/gpr/alloc.h',
                               'src/core/lib/gpr/env.h',
