+++ package.xml
@@ -770,7 +770,7 @@
     <file baseinstalldir="/" name="src/core/lib/event_engine/endpoint_config.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/event_engine/endpoint_config_internal.h" role="src" />
     <file baseinstalldir="/" name="src/core/lib/event_engine/event_engine.cc" role="src" />
-    <file baseinstalldir="/" name="src/core/lib/event_engine/event_engine_internal.h" role="src" />
+    <file baseinstalldir="/" name="src/core/lib/event_engine/event_engine_factory.h" role="src" />
     <file baseinstalldir="/" name="src/core/lib/event_engine/sockaddr.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/event_engine/sockaddr.h" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gpr/alloc.cc" role="src" />
