+++ tools/doxygen/Doxyfile.c++.internal
@@ -1723,7 +1723,7 @@ src/core/lib/debug/trace.h \
 src/core/lib/event_engine/endpoint_config.cc \
 src/core/lib/event_engine/endpoint_config_internal.h \
 src/core/lib/event_engine/event_engine.cc \
-src/core/lib/event_engine/event_engine_internal.h \
+src/core/lib/event_engine/event_engine_factory.h \
 src/core/lib/event_engine/sockaddr.cc \
 src/core/lib/event_engine/sockaddr.h \
 src/core/lib/gpr/alloc.cc \
